Make the Finalizing state equal to Complete
In the logic there is a case where Finalizing state
of Velero backup is considered Complete and PVCs are
allowed to be mounted, this is not reflected in the
State, which transitions VMFR to a terminal phase.

The terminal phase won't allow state updates, so
VMFR is left with Finalizing status.

Signed-off-by: Michal Pryc <mpryc@redhat.com>